Build docs issues fixes (#3075)
* delayed imports

* move parse density to parse composition

* line deletion
